Scale nexrad_radial_get_*() azimuths to <>360
    * Modify nexrad_radial_get_rangebin() to accept ints, not uint16_ts

    * Modify nexrad_radial_get_rangebin(), nexrad_radial_get_ray() to scale
      azimuth inputs to <> 360 range

+++ include/nexrad/radial.h
@@ -66,8 +66,8 @@ nexrad_radial_ray *nexrad_radial_get_ray(nexrad_radial *radial,
 int nexrad_radial_ray_get_azimuth(nexrad_radial_ray *ray);
 
 int nexrad_radial_get_rangebin(nexrad_radial *radial,
-    uint16_t azimuth,
-    uint16_t range
+    int azimuth,
+    int range
 );
 
 nexrad_radial_ray *nexrad_radial_read_ray(nexrad_radial *radial,
